seperate out homepage preview components
