+++ backend/tsconfig.json
@@ -15,7 +15,7 @@
     "noEmit": true,
 
     // Best practices
-    "strict": true,
+    "strict": false, //smd 
     "skipLibCheck": true,
     "noFallthroughCasesInSwitch": true,
