+++ frontend/src/app/events/page.tsx
@@ -11,7 +11,7 @@ import Marquee from "@/components/ui/marquee";
 import "@/components/calendar/Calendar.css";
 import { motion } from "framer-motion";
 import { config } from "@/config";
-import { LampContainer } from "@/components/lamp";
+import { LampContainer } from "@/components/ui/lamp";
 import { Event } from "@/components/calendar/types";
 
 export default function Events() {
@@ -139,7 +139,7 @@ const EventCard = ({ event }: { event: Event }) => (
                    shadow-xl hover:shadow-2xl transition-all duration-300 
                    w-[300px] h-[400px]"
   >
-    Image Section
+    {/* Image Section */}
     <div className="h-[160px] flex-shrink-0">
       {event.cover_image_url ? (
         <img
