+++ frontend/src/components/ui/illustration.tsx
