+++ frontend/src/components/ui/lamp.tsx
